removed redundant code; fixed pav plotting; added documentation
