+++ lir/__init__.py
@@ -1,7 +1,7 @@
+from . import metrics
 from .lr import *
 from .bayeserror import *
 from .calibration import *
 from .ece import *
 from .generators import *
-from .pav import *
 from .util import Xn_to_Xy, Xy_to_Xn
