+++ setup.py
@@ -16,7 +16,7 @@ setup(name='lir',
       url='https://github.com/HolmesNL/lir/',
       author='Netherlands Forensic Institute',
       author_email='fbda@holmes.nl',
-      packages=[os.path.join(package_dir,'lir')],
+      packages=['lir'],
       setup_requires=['nose'],
       test_suite='nose.collector',
       install_requires=packages,
